% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_regression_ols}
\alias{linear_regression_ols}
\title{Linear Regression OLS}
\usage{
linear_regression_ols(formula, y, data)
}
\arguments{
\item{formula}{formula: describes the model to be fitted symbolically.}

\item{y}{character denoting the name of the column in data to be used as the response variable.}

\item{data}{data frame of data to be used.}
}
\value{
Returns estimates of the coefficients from a linear regression model, including their standard errors, t-statistics, and 2-sided p-values. This function also automatically outputs additional information such as a five number summary of the residuals, R-square and adjusted R-squared values, and an F-test,
}
\description{
Estimates the coefficients of a linear model, including their standard errors, t-statistic, and corresponding p-value.
}
\examples{
library(lineRegOLS)
set.seed(202311)
score = rnorm(1000, mean = 10, sd = 3)

age = rnorm(1000, mean = 35, sd = 7.5)

perf = round(runif(1000, 0, 2))
perf[perf == 0] = "low"
perf[perf == 1] = "medium"
perf[perf == 2] = "high"

gender = round(runif(1000, 0, 1))
gender[gender == "0"] = "male"
gender[gender == "1"] = "female"

testNum = round(runif(1000, 0, 2))
testNum[testNum == 0] = "one"
testNum[testNum == 1] = "two"
testNum[testNum == 2] = "three"

testScores = data.frame(score, age, perf, gender, testNum)

linear_regression_ols(score~age+gender+perf+testNum, y = "score", data = testScores)
linear_regression_ols(score~-1+age+gender+perf+testNum, y = "score", data = testScores)

}
