% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_model.R
\name{linear_regression_ols}
\alias{linear_regression_ols}
\title{Linear Regression OLS}
\usage{
linear_regression_ols(
  data,
  y,
  x,
  cat_coding = "reference",
  reference_cat = NULL
)
}
\arguments{
\item{data}{Data frame of data used.}

\item{y}{Character denoting the name of the column in data to be used as the response variable.}

\item{x}{Vector of characters denoting the columns in data to be used as the predictor variables. Can handle categorical variables.}

\item{cat_coding}{Argument to set coding scheme to use in place of non-binary categorical covariates. Defaults to "reference" cell coding. Enter "means" for mean cell coding.}

\item{reference_cat}{String argument to set reference category for categorical covariates. Defaults to first category in multi-level categorical covariates.}
}
\value{
Estimates of the linear regression coefficients, their standard errors, t-statistics, and two-sided p-values.
}
\description{
Estimates the coefficients of a linear model, including their standard errors, t-statistic, and corresponding p-value.
}
\examples{
set.seed(202311)
score = rnorm(1000, mean = 10, sd = 3)

age = rnorm(1000, mean = 35, sd = 7.5)

perf = round(runif(1000, 0, 2))
perf[perf == 0] = "low"
perf[perf == 1] = "medium"
perf[perf == 2] = "high"

gender = round(runif(1000, 0, 1))
gender[gender == "0"] = "male"
gender[gender == "1"] = "female"

testGrades = data.frame(score, age, perf, gender)

linear_regression_ols(data = testData, y = "score", x = c("age", "perf", "gender"), cat_coding = "means", reference_cat = NULL)
linear_regression_ols(data = testData, y = "score", x = c("age", "perf", "gender"), reference_cat = "low")

}
